Reorganise pgdo-cli, move commands into modules
