+++ pgdo-cli/src/command.rs
@@ -1,0 +1,3 @@
+pub(crate) mod exec;
+pub(crate) mod runtimes;
+pub(crate) mod shell;
