+++ pgdo-cli/src/command/shell.rs
@@ -1,0 +1,37 @@
+use color_eyre::eyre::{Result, WrapErr};
+
+use crate::{args, runner};
+
+#[derive(clap::Args)]
+#[clap(next_help_heading = Some("Options for shell"))]
+pub struct Args {
+    #[clap(flatten)]
+    pub cluster: args::ClusterArgs,
+
+    #[clap(flatten)]
+    pub database: args::DatabaseArgs,
+
+    #[clap(flatten)]
+    pub lifecycle: args::LifecycleArgs,
+
+    #[clap(flatten)]
+    pub runtime: args::RuntimeArgs,
+}
+
+pub fn invoke(args: Args) -> Result<i32> {
+    let Args { cluster, database, lifecycle, runtime } = args;
+    runner::run(
+        cluster.dir,
+        &database.name,
+        runner::determine_strategy(runtime.fallback)?,
+        lifecycle.destroy,
+        runner::initialise(cluster.mode),
+        |cluster| {
+            runner::check_exit(
+                cluster
+                    .shell(&database.name)
+                    .wrap_err("Starting PostgreSQL shell in cluster failed")?,
+            )
+        },
+    )
+}
